fix: when sorting, allow doclets with different scopes to be interleaved
Previously, if you had 20 doclets with the same `memberof`, but 5 had the scope `inner` and the other 15 had the scope `static`, the 15 `static` doclets always came before the 5 `inner` doclets.

Now, doclets are sorted on their `name` first, so doclets with the same `memberof` but different scopes can be interleaved.