refactor: tidy up promise-handling in tests
In some cases, we check that a return value is a `Promise`, so we can't just `await` it. That means we need to handle the promise afterwards. In these cases, we typically just want to eat the promise; we don't care how it's fulfilled or rejected.

Now we have a helper method to do that more succinctly.